fix import cycle due to acmock usage in internal test
